Post-validation module extracted.
New `Post#extend_with_validation` instance added to make intention
explicit.

RSpec: 732 examples, 0 failures, 1 pending
RuboCop: 241 files inspected, no offenses detected
