+++ app/entities/post.rb
@@ -1,8 +1,6 @@
 
-# require_relative 'post/byline_builder'
-# require_relative 'post/image_body_builder'
-# require_relative 'post/text_body_builder'
 require_relative 'post/extensions/presentation'
+require_relative 'post/extensions/validation'
 
 # Namespace containing all application-defined entities.
 module Entity
@@ -15,6 +13,10 @@ module Entity
 
     def extend_with_presentation
       extend Extensions::Presentation
+    end
+
+    def extend_with_validation
+      extend Extensions::Validation
     end
 
     def initialize(attributes)
@@ -35,34 +37,11 @@ module Entity
       attributes[:slug].present?
     end
 
-    def valid?
-      valid_title? && valid_author_name? && body_or_image_post?
-    end
-
     private
 
     def attrib_keys
       [:author_name, :body, :created_at, :image_url, :pubdate, :slug, :title,
        :updated_at]
-    end
-
-    def body_or_image_post?
-      return true if attributes[:body].to_s.strip.present?
-      attributes[:image_url].to_s.strip.present?
-    end
-
-    def registered_author?
-      attributes[:author_name] != 'Guest User'
-    end
-
-    def valid_author_name?
-      name = attributes[:author_name]
-      name.present? && name == name.strip && registered_author?
-    end
-
-    def valid_title?
-      title = attributes[:title]
-      title.present? && title == title.strip
     end
   end # class Entity::Post
 end
